+++ src/ast/lists/declaratorList.cpp
@@ -4,7 +4,7 @@
 void DeclaratorList::printC(std::ostream &os) const
 {
     std::vector<StatementPtr> initializations = getNodes();
-    for (int i = 0; i < initializations.size(); ++i)
+    for (size_t i = 0; i < initializations.size(); ++i)
     {
         if (i > 0) { os << ", "; }
         os << initializations[i];
@@ -23,7 +23,7 @@ int DeclaratorList::getDeclCount() const
 
 void DeclaratorList::generatePython(std::ostream &os, PythonContext &context, int scopeDepth) const
 {
-    for (int i = 0; i < branches.size(); ++i)
+    for (size_t i = 0; i < branches.size(); ++i)
     {
         DeclaratorPtr decl = getDeclarator(i);
         if (i > 0) { context.indentStream(os, scopeDepth); }
@@ -37,7 +37,7 @@ void DeclaratorList::generatePython(std::ostream &os, PythonContext &context, in
 
 void DeclaratorList::generateIL(std::vector<Instr> &instrs, ILContext &context, std::string destReg) const
 {
-    for (int i = 0; i < branches.size(); ++i)
+    for (size_t i = 0; i < branches.size(); ++i)
     {
         getDeclarator(i)->generateIL(instrs, context, destReg);
     }
