+++ src/codegen/allocationFrame.cpp
@@ -3,7 +3,7 @@
 
 void AllocationFrame::allocate(Allocation allocation)
 {
-    for (int i = 0; i < _allocations.size(); ++i) { _allocationOffsets[i] += allocation.size; }
+    for (size_t i = 0; i < _allocations.size(); ++i) { _allocationOffsets[i] += allocation.size; }
     _allocations.push_back(allocation);
     _allocationOffsets.push_back(0);
     _frameSize += allocation.size;
@@ -28,7 +28,7 @@ int AllocationFrame::getFrameSize() const
 
 int AllocationFrame::getAllocationPosition(std::string name) const
 {
-    for (int i = _allocations.size() - 1; i >= 0; --i)
+    for (size_t i = _allocations.size() - 1; i >= 0; --i)
     {
         if (name == _allocations[i].name) { return i; }
     }
